canonical_documentation/requirements.txt
pyyaml
nox